requirements.txt
flask==3.0.3
tensorflow-cpu==2.16.1
numpy==1.26.4
opencv-python==4.10.0.84
flask-cors==4.0.1
gunicorn==22.0.0
gdown==5.2.0